Refactor code by separating about window and folder logic
